requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.0
colorama==0.4.6
deno==2.6.4
filelock==3.20.2
fsspec==2025.12.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
psutil==7.2.1
pycparser==3.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
selenium==4.40.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
speedtest-cli==2.1.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
